Fixed fetch commands for user
Finally managed to get the fetch active commands for users. This will need to be advanced in such a way that globally commands can be managed without affecting local. Apart from, development and deprecated users stand free to enable/ disable commands on their own. This will be an interesting feature to manage.