Improvements on dashboard and UI/UX
I've fixed elements on the user and admin dashboard, improved UI/UX for siderbar and adminsidebar as-well as fixed and re-created some of the dashboard features. It's time to dwelve into more complicated tasks of the application, therefore I have started on the command creation module along with more advanced command and feature management within the dashboard itself. Somewhere along the next steps would be a good point to introduce the rest of the team to the tool.